feat(website): add an og-what card and share one card shell
/what-is-webjs had no card of its own and unfurled with the site-wide
og.png, whose headline is the product tagline. That page exists to answer
one question (its title, h1 and slug all match it, a definition sits in
the first 160 characters, and a visible FAQ backs FAQPage JSON-LD), so it
is the likeliest of the three to be shared AS an answer, and it was the
one arriving under a card that answered something else.

Adding it would have made three near-identical generators, which is the
count at which a fix stops landing on every card. That is not
hypothetical: it is how the /why card kept a retired logo and a claim the
site had already corrected. So the shell moved to scripts/lib/og-card.mjs
(palette, inlined faces, lockup, top row, panels, footer, render, fit
pass) and each card now owns only its content. It is deliberately not a
card framework: a card that wants a different structure passes its own
CSS.

The two fact panels are shared the same way, and the home card adopts
them. It was a headline capped at 19ch over a lede at 34ch, so the right
half of a 1200px card was empty; the panels give it the composition the
/why card already had. The kicker also moves into the shell, so all three
carry BUILT FOR THE AI ERA opposite the lockup rather than one of them
having it.

The /why card is visually unchanged by the refactor.
